requirements.txt
# Core dependencies
asyncpg>=0.29.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0

# Development
black>=24.0.0
mypy>=1.8.0
